updating to compare
