+++ test/unit/test_events.py
@@ -1,7 +1,8 @@
 from dbt.events import AdapterLogger
 from dbt.events.types import AdapterEventDebug
+from dbt.events.base_types import Cli,File
+import inspect
 from unittest import TestCase
-
 
 class TestAdapterLogger(TestCase):
 
@@ -45,5 +46,25 @@ class TestAdapterLogger(TestCase):
         # were no args passed after the initial message
         event = AdapterEventDebug(name="dbt_tests", base_msg="boop{x}boop", args=())
         self.assertTrue("boop{x}boop" in event.message())
+
+class TestEventCodes(TestCase):
+
+    def setUp(self):
+        pass
+
+    # this interface is documented for adapter maintainers to plug into
+    # so we should test that it at the very least doesn't explode.
+    def test_event_codes(self):
+        all_concrete = set(Cli.__subclasses__()) \
+            .union(set(File.__subclasses__()))
+        all_codes = {}
+
+        for event in all_concrete:
+            if not inspect.isabstract(event):
+                # must be in the form 1 capital letter, 3 digits
+                self.assertTrue('^[A-Z][0-9]{3}', event.code())
+                # cannot have been used already
+                self.assertFalse(event.code() in all_codes)
+                all_codes.add(event.code())
